Move the plugin installation logic in a dedicated class
